API

In the following, you find the documentation of all exported functions of the BartIO.jl package:

BartIO.bart

BartIO.read_cfl

BartIO.set_bart_path

BartIO.write_cfl

BartIO.bart — Method

bart = wrapper_bart(pathtobart::String)
Args are concatenated at the end of the command.

### output
- bart : a wrapper to call bart from Julia through the python functions from the bart repository.

TO DO :
- add kwargs support like in the python wrapper https://github.com/mrirecon/bart/pull/295

Example :
````
bart = wrapper_bart(pathtobart)
bart(0,"version")

bart(0,"phantom -h")
bart(1,"phantom -k -x128")
````

BartIO.read_cfl — Method

read_cfl(filename::String)

read_cfl(filename(no extension)) -> Array{ComplexF32,N} where N is defined the filename.hdr file

read_cfl(filename.cfl) -> Array{ComplexF32,N} where N is defined the filename.hdr file

read_cfl(filename.hdr) -> Array{ComplexF32,N} where N is defined the filename.hdr file

Reads complex data from files created by the Berkeley Advanced Reconstruction Toolbox (BART). The output is an Array of ComplexF32 with the dimensions stored in a .hdr file.

Parameters:

filename: path and filename of the cfl and hdr files, which can either be without extension, end on .cfl, or end on .hdr

BartIO.set_bart_path — Method

set_bart_path(pathToBart::String)

Define the path to the BART toolbox (store in ENV["TOOLBOX_PATH"])

BartIO.write_cfl — Method

write_cfl(filename::String,dataCfl::Array{ComplexF32})

write_cfl(filename(no extension),Array{ComplexF32})

write_cfl(filename.cfl, Array{ComplexF32})

write_cfl(filename.hdr,Array{ComplexF32})

Write complex data to files following the convention of the Berkeley Advanced Reconstruction Toolbox (BART). The input is an Array of ComplexF32 with the dimensions stored in a .hdr file.

Parameters:

filename: path and filename of the cfl and hdr files, which can either be without extension, end on .cfl, or end on .hdr

Array{ComplexF32,N}: Array of ComplexF32 corresponding to image/k-space